Cleanup start logic into gui module
