+++ setup.cfg
@@ -19,7 +19,7 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    usdb_syncer = usdb_syncer.main:cli_entry
+    usdb_syncer = usdb_syncer.gui:main
     generate_pyside_files = tools.generate_pyside_files:cli_entry
     generate_song_list_json = tools.generate_song_list_json:cli_entry
     write_release_info = tools.write_release_info:cli_entry
