+++ src/usdb_syncer/gui/search_tree/tree.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Callable
 from PySide6.QtCore import QModelIndex, Qt
 
 from usdb_syncer import db, events
-from usdb_syncer.gui.utils import keyboard_modifiers
+from usdb_syncer.gui.gui_utils import keyboard_modifiers
 
 from .item import Filter, FilterItem, RootItem, VariantItem
 from .model import TreeModel, TreeProxyModel
